Rename ReactContext.h to MsoReactContext.h because there is another file of the same name which leads to issues on other build systems (e.g. Buck/Clang)
