+++ vnext/Microsoft.ReactNative/ReactHost/MsoReactContext.h
